requirements.txt
opencv-python
pillow
openai
